refactor event hook directory
